+++ src/bin/oxide.rs
@@ -4,6 +4,8 @@ use peroxide::*;
 
 fn main() {
     let a = matrix!(1;4;1, 2, 2, Row);
-    let (p,l,u) = a.plu();
-    println!("{}\n{}\n{}", p, l ,u);
+    let pqlu = a.lu().unwrap();
+    println!("{:?}\n{:?}\n{}\n{}", pqlu.p, pqlu.q, pqlu.l, pqlu.u);
+    println!("{}", a.det());
+    println!("{}", a.inv().unwrap());
 }